backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
celery[redis]==5.4.0
redis==5.2.1
channels==4.2.0
channels-redis==4.2.1
langchain==0.3.20
langchain-openai==0.3.9
langchain-tavily==0.2.4
tavily-python==0.5.0
langgraph==0.3.21
openai==1.66.3
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
dj-database-url==2.3.0
whitenoise==6.8.2
Pillow==11.1.0
markdown==3.7
bleach==6.2.0
djangorestframework-simplejwt==5.4.0
